requirements.txt
# Core dependencies
torch>=2.5.0
transformers>=4.57.0
datasets>=2.0.0
tokenizers>=0.15.0

# Evaluation metrics
nltk>=3.8
rouge-score>=0.1.2
evaluate>=0.4.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
tabulate>=0.9.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.64.0
psutil>=5.9.0

# Testing
pytest>=7.0.0
hypothesis>=6.0.0

# Development
black>=22.0.0
flake8>=5.0.0